Agentic AI/Pratice Assignment -01/open_router_1/openai_agent_project/pyproject.toml
[project]
name = "openai-agent-project"
version = "0.1.0"
description = "An agent using OpenAI Agents SDK"
readme = "README.md"
authors = [
    { name = "nidanasar", email = "nidashafqet@gmail.com" }
]
requires-python = ">=3.11"
dependencies = ["openai", "agents", "python-dotenv"]

[project.scripts]
openai-agent-project = "openai_agent_project.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
